ComputerUseAgent/GUI_sandbox/client/requirements.txt
requests==2.26.0
pillow==9.0.1
argparse==1.4.0
